修了下界面

+++ app/src/main/java/com/online/factory/factoryonline/modules/agent/AgentViewModel.java
@@ -69,12 +69,12 @@ public class AgentViewModel extends BaseObservable {
 
     @Bindable
     public String getArea() {
-        return factory.getRange()+"㎡";
+        return (int)factory.getRange()+"㎡";
     }
 
     @Bindable
     public String getTotalPrice() {
-        return factory.getPrice() * factory.getRange() + "元/月";
+        return (int)(factory.getPrice() * factory.getRange()) + "元/月";
     }
 
     @Bindable
